+++ src/avr/src/drv8711.h
@@ -173,7 +173,6 @@ typedef void (*stall_callback_t)(int driver);
 
 
 void drv8711_init();
-void drv8711_shutdown();
 drv8711_state_t drv8711_get_state(int driver);
 void drv8711_set_state(int driver, drv8711_state_t state);
 void drv8711_set_microsteps(int driver, uint16_t msteps);